reporter: Move most logic from `NoticeReporter` to an abstract class
Only keep the logic to find the licenses to add to the notice file in the
`NoticeReporter` class. This allows to add a notice reporter that uses a
different filter for licenses in a later commit.
